requirements-extract-local.txt
boto3==1.35.29
botocore==1.35.29
python-dotenv==1.0.1
oracledb==2.4.1
